Extract ProductImage related endpoints into a separate controller
