+++ Catalog/Catalog.Api/Controllers/v1/Abstractions/IProductController.cs
@@ -10,8 +10,4 @@ public interface IProductController
     Task<IActionResult> AddProduct(ProductWriteDto productDto);
     Task<IActionResult> UpdateProduct(Guid id, ProductWriteDto productDto);
     Task<IActionResult> DeleteProduct(Guid id);
-    Task<IActionResult> GetAllProductImagesByProductId(Guid productId);
-    Task<IActionResult> AddProductImage(Guid productId, ProductImageWriteDto productImageDto);
-    Task<IActionResult> UpdateProductImage(Guid id, ProductImageWriteDto productImageDto);
-    Task<IActionResult> DeleteProductImage(Guid id);
 }
